I added code to the MATLAB scatter/contour plots to handle the problem of a minimum contour for the CLUBB bivariate joint PDF contour plot.
The problem was that MATLAB automatically selected contours based on a number
set by the user.  For some plots where one component had a very large spread in
both variables and a very small weight (mixture fraction), it would take a ton
of lines to resolve any line for one of the components.  An example of this was
my RICO plots for chi and eta.  There is a small mixture fraction, and the 1st
PDF component had huge spreads for both chi and eta.  This resulted in a very
small P(chi,eta) contribution from component 1 compared to component 2.  It
took 2500 contour lines to resolve a single line circling the 1st PDF component.

It order to handle this problem, I used the MATLAB option to pass in a contour
vector instead of a number of lines.  The contour vector is user-specified and
must be monotonically increasing.  This allowed me to set a small, non-zero
value for the lowest contour, which would allow me to resolve some part of the
dwarfed component without a ton of contour lines.

I needed a method to determine the lowest contour.  The method is rather
complicated.  Basically, I pick a point for both PDF components.  This point
is a specified number of standard deviations away (in each variable) from the
bivariate mean for each PDF component.  The value of P(x,y) is calculated at
both of these points.  The minimum value of P(x,y) between these two points is
the minimum contour.

In the event that there is so much spread to one of the components that the
contour would fall outside of the domain of the plot, I have a backup method.
The backup method sorts all values of P(x,y) over the domain of the plot, and
arranges them in a vector from smallest to largest.  The 10th percentile value
of P(x,y) (larger than only 10% of the calculated P(x,y) values) becomes the
value of the lowest contour.

The highest contour is determined by the maximum calculated P(x,y).  The number
of lines beteween the two contours depends on how many contours the user wants
to use.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@7330 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: f54a15ef0448736d211c1a892d39a030be6c6c55