I am implementing a new `G_unit_tests` test for the new SILHS category chooser.
This unit test picks 100 sample points. It picks the probabilities of each of eight categories such that some number of sample points are expected to fall in each category. This is then verified to be the case.

Note that even though random numbers are in play, it is theoretically guaranteed that the correct number of sample points will fall into each box, because the determination variate is stratified (in the test and in the real SILHS).

I highly doubt that output will be bin-diff unidentical in any configuration.

This is for clubb:ticket:736.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@7356 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: a5ff0b46c52db4477cf24da44ac2176e462b60eb