Happy Thanksgiving to all!!!
We all have plenty in our lives to be thankful for.

One additional little thing that we can now be thankful for is that the input
fields code is fixed.  It can now (once again) apply input fields correctly
when the LES input file does not start at 00:00.

This was touched upon in clubb:ticket:735.

Enjoy stuffing yourselves with stuffing!



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@7411 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 3acba8a0dba7fba863265cde84892bb5dc71af7f