clubb:ticket:762
I spent some time today debugging the PGI compile test failure. The error was also reproducible on `gfortran` when running in single precision.

I'm not yet completely convinced that this is the problem, but my best guess is that the value of `prob_thresh` currently set in the prescribed-probability code is too low and causes too much stress on single-precision CLUBB. The value was set at `prob_thresh = 1.0e-8`. When cloud fraction is on the order of, e.g., 0.999999, single precision CLUBB produces an inaccurate value of cloud fraction and other variables, causing the assertion check to trigger.

Also, I don't claim to know anything about atmospheric science, but 1.0e-8 seems unphysically low.

Therefore, I set the value to 5.0e-3 in this commit. Now, all the SILHS cases I tested run to completion using the PGI compiler. I'm not yet sure if this is all we need to do; I'll investigate some more at some point.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@7477 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 93e9120b72ad46950a8751f8976ea279d6ace3ab