I think I have finished my category variance sampler. This new module computes the variance of any microphysics output in the eight SILHS importance categories at every height level.
The output is not turned on by default, and therefore output is bin-diff identical. When one turns on the new output (by adding 'silhs_variance_category' to the `stats.in` file), output is bin-diff identical except for the `lh_zt.*` file(s), where the new variables are included.

See clubb:ticket:752.

----

As an aside, I am proud to announce that this is my 500th commit to the CLUBB repository!! (Technically, it is my 501st commit, but r7064 was committed as "raut" instead of "raut@uwm.edu", so that doesn't count.) If I make it to 1000 commits before I depart from the group, I can cross an item off of my bucket list! :)



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@7494 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: e298f2a343fb15be82982093aca3d09e05cab0b5