I am adding the code for the new scheme to calculate mu_hm_1 and mu_hm_2 for each hydrometeor species.  The scheme preserves the overall mean and the overall variance of the hydrometeor.
Currently, this new code is not hooked up to the rest of the CLUBB code.  It
is just placed here to get it in the code and compile with the rest of the code.

The next step is to hook it up to G-unit tests!

Personal CLUBB commit counter:  1151 (sponsered in part by clubb:ticket:755).



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@7504 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: fab92d897c8c604d4f431191154775b1351bd3c5