I am adding the tunable parameters for my new omicron and zeta variables to the CLUBB trunk (and again, blah ... blah ... blah ... to my special input fields code where they will be of great use).
Since the new code is still only connected to the G-unit tests, this changes doesn't
affect results now.

The results in the GrADS output files from this revision are an exact match to the
results in the GrADS output files from the previous revision for all cases.

Personal CLUBB commit counter:  1153 (sponsored in part by clubb:ticket:755).



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@7506 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: a69a6e0b93bfe2ce1931da35b583f353c6679046