I made a new subroutine out of the code that calculates component precipitation fraction based on weighted cloud_frac_1 above.  Now, each main method used to calculate component precipitation fraction has its own subroutine.
The results in GrADS output files from THIS PORTION of this revision are an
exact match to the results in GrADS output files from the previous revison for
all cases.

However, along the way I noticed a bug that I had introduced in the previous
revision of this code.  I fixed that bug.  Hopefully, this bug is the reason the
old "ratio method" acted so woefully!

Trac System Reference dedicated to glorious March days: clubb:ticket:755.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@7549 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 94bb7258fd22dc23a0ce693600fad4be24f912ec