This is the first step in updating the scatter/contour plotter to be able to plot multiple CLUBB lines on one plot!
This will require vectorizing many places in the scatter/contour plotter code.

However, in this first step, the run_scatter_contour_plots.bash script is
updated to be able to read in multiple CLUBB data sets (files).

As before, if less than two files (and paths) are given at the command line,
the script reverts to the "default" configuration listed in the
PDF_scatter_contour_plotter.m file, which is a specific RICO LES 3D output file
and one specific CLUBB RICO output file.

However, if at least two files (and paths) are given at the command line, the
script now has an enhancement.  Just as before, the first filename listed is the
SAM LES 3D output file.  However, after that, one or more CLUBB files can be
listed at the command line.  The script will read in the filenames into an array,
print them to the screen, and then pass them into the MATLAB code.  As of now,
only the first CLUBB file listed is used in the actual program.  It will remain
that way until the entire program has been updated.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@7700 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 898015bc34828d4d467310a6717d0242985165fc