I am continuing my quest to create an automated script to generate plots for the paper!
In this commit, to avoid creating giant output data sets, I added a `stats.in` file that specifies the small number of statistics that are actually needed for the plots that will be put in the paper. I also modified the SILHS run script to accept an option for the statistics file.

This is for clubb:ticket:740.


git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@7712 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: ea740dabd0b3b4cdc06d58ce883cde68e076c053