I improved the organization and functionality of the scatter/contour plotter.
When the program is run, a new directory, named by the date, is created and
the new output is placed in it.  This directory can be easily renamed.

I also added an option to enter a descriptive addition to the filenames.

In order to make the ln rr and ln Nr plots easier to read, I moved the legend
from the upper-right corner to the upper-left corner.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@7764 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 20517ba1f1fa68ec9a847771b5a4eda32255812b