I am adding a function and file to the scatter/contour plotter.  This file is made out of the code that reads the CLUBB files and then checks them for consistency with one another.  This will make it easier to call the necessary functions from the upcoming matlab_goodness_of_fit_tests directory.
The .png output files from this portion of this revision are an exact match to
the .png output files from the previous revision.

I also changed the lower bound of the log-scaled marginal plots for w, rt,
theta-l, chi, and eta, matching a previous change to the log-scaled marginal
plots for rr and Nr.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@7767 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 027d031a96ae19f8f79cf6141a51b9ffdbe988ac