For clubb:tickets:770, I am merging my branch into the trunk. This code includes a 3D_Luhar pdf closure and is provided 'as is' and without warranty.
I tested the priority cases from when the branch was created to the merged copy. The differences are very small and could be a result of subtle numerical changes such as a change in order of operations. When overplotted, the differences are not discernable.

ADG2 does not nessesarily work for interactive cases since CLUBB is mostly written assuming ADG1. Hence, I tested arm_97. For arm_97, the differences were very tiny (similar to ADG1) and when overplotted not discernable.

The code could use some refactoring. However, it is usable in case others would like to experiment (comment:68:ticket:764)

git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@7802 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 586c13884d8e6def94699a6b12679e977f1bbe83