I have discovered a better way to do the assertion check for consistency between `chi_i`, `stdev_chi_i`, and `cloud_frac_i`.
In this commit, I create a (very) small box around cloud_frac_i. (1-cloud_frac_i should be equal to zero.) I then verify that the left end of the box gives a positive value of chi in normal space (that is, in cloud), and that the right end of the box gives a zero or negative value of chi in normal space (that is, no cloud).

This is better than how it was done before because the tolerance (which is the box width) has a more meaningful interpretation. E.g., if a box of width 10^-5^ passes the assertion check, then it is guaranteed that all uniform values of chi outside of this box (i.e., where 1 - 10^-5^ of points are expected to fall) will be correctly interepreted as in cloud or out of cloud. Before, I had to set something called `boundary_tol`, which I lamented that I didn't know how to set properly.

This is for clubb:ticket:720!



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@7891 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 527b33a09df1fee8e2ed45a2597a459706c4db27