I added a script called run_DDL_DL_SL_cases.bash to the run_scripts directory in the Brian special input fields branch.  This script compiles and runs the code for RICO, DYCOMS-II RF02 DS, and LBA using the default DDL hydrometeor PDF shape.  Then, it sets up the DL hydrometeor PDF shape and runs the same cases. After that, it sets up the SL hydrometeor PDF shape, recompiles the code, and runs the same cases again.  All the results (in netCDF output) are stored in the output directory in a subdirectory that is named by a date string.  This directory contains subdirectories named "DDL", "DL", and "SL" that store the appropriate results.  After all this is complete, the script reverts the changes made for the DL and SL hydrometeor PDF shapes, restoring the code to its original state.
The results in netCDF output files produced by this script are an exact match
to the results in netCDF output files produced by my previous efforts for all
cases (RICO, DYCOMS-II RF02 DS, and LBA) and for all hydrometeor PDF types (DDL,
DL, and SL).

The "I've never referenced the Papers Trac in a commit before" Trac System
Reference:  papers:ticket:7.  It seems like it's more relevant because we want
to reproduce results for a paper with ease.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/branches/Brian_special_input_fields_r7465@7902 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 59b51b8bba486b3f7e80c62bd77b0fd8ac1b53c2