I finished the option to run the goodness-of-fit tests as part of the run_DDL_DL_SL_cases.bash script.  The results of the goodness-of-fit tests are an exact match to those produced by hand-running the test scripts back on September 3, 2015.
In all, the script can now run each of RICO, DYCOMS-II RF02 DS, and LBA for each
of DDL, DL, and SL hydrometeor PDFs and produce results that exactly match the
previous benchmarks.  Then, it can produce the PDF plots, run the
goodness-of-fit tests, and produce the profiles of warm microphysics fields, all
with results that exactly match the benchmark plots, tests, and profiles that
were previously made. 



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/branches/Brian_special_input_fields_r7465@7908 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: c287b4c6d421ef96a00b01062373e32090710a80