For clubb:ticket:779:
I am officially upgrading CLUBB standalone to be Fortran 2003 compliant! As a test, I added an allocatable array to a test derived type in `clubb_driver.F90` and compiled it with all of our compilers. This change to the `gfortran` configuration is the only change I had to make!

Compiler status:
* gfortran -- success!
* ifort -- success!
* pgi -- success!
* oracle -- success!
* g95 -- unrelated error about not finding "-lf77blas"
 * This made me wonder how we compile using g95 in the nightly Bitten tests. As it turns out, we don't! The description of the "Nightly Build (External Users)" configuration says it uses G95, but it actually uses gfortran. I guess g95 is kind of dead.
* open64 -- success!


git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@7985 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 6b5db26c86960668e3313f6235f1c07eebebb36e