For clubb:ticket:782:
I merged in NCAR's changes to `advance_clubb_core_module.F90` and `interpolation.F90`. This is probably at least half the battle.

I added some flags to `model_flags.F90`. They are set to the values that preserve answers in CLUBB standalone. Output is bin-diff identical for a comprehensive selection of test cases.


git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@7989 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 09326d9dfe1aaff864a12cd1c7c0773002d912f4