For clubb:ticket:753:
I refactored the code in `lh_microphys_var_covar_module` to compute the SILHS variance and covariance tendencies by using the terms given in clubb:ticket:753#comment:20. This is the first step in making SILHS agree with the analytic solution.

I overplotted the tendencies produced by this code with the tendencies produced by the prior code, and visually they look identical. Nevertheless, output was not bin-diff identical (likely due to round-off error).



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@7998 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 8b95edc889d65111b47e8e4314e7bfbde748f650