For clubb:ticket:753:
I added a flag, `l_lh_instant_var_covar_src`, which will include the timestep-dependent terms when false, and set the timestep-dependent terms to zero when true.

The flag is turned off by default, so the output from this revision is bin-diff identical to the output from the __previous__ revision (r7998).


git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@7999 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 39f1cb066a06de34f9324db0c1755e66ba28bf2a