For clubb:ticket:786:
I fixed two tuner related problems.

I made the SILHS parameters namelist optional, so that a warning message is simply printed if the namelist is not found. I'm not particularly fond of this solution, but the difficult alternative is to track down the places in the tuner where the SILHS namelist needs to be placed. Maybe some other day.

I also deallocated a variable in SILHS that wasn't deallocated before.

Output ought not to change.


git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@8023 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: ccef9fe7bcd6531d3a6903fbe2c6a8db763830ea