For cam:ticket:56:
While looking through plots, I noticed that for some reason, CLUBB output files list `cloud_frac` as having units of "count", which in fact it is a fraction between 0 and 1. I have changed the units to be "-", as is done elsewhere.

Note that this commit is BIT_CHANGING, but only because of the metadata change. Numerical values of all variables should be identical.


git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@8039 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 386dd79390ff0720658e567be846b45f1267b596