For cam:ticket:59:
While reading through the code, I found an important bug in the moist Brunt-Vaisala code. Inside a `where` statement, I had calls to functions that take in an entire profile as input, such as `zt2zm` and `ddzt`. But the arguments to these functions where themselves initialized in the `where` statement! Therefore, much of the input profiles to these functions were likely uninitialized. Not good.

This change affects only the case where `l_brunt_vaisala_freq_moist = T`, which is not the default.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@8085 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 8ca078a9223a3a07a90e115f1f1d3296ea196896