For cam:ticket:73:
clubb:r8220 produced bin-diff identical ''GrADS'' files, but not bin-diff identical NetCDF files because CLUBB outputs all tuning parameters to the NetCDF files, and a tuning parameter was added in the aforementioned changeset. This caused the Nightly Bin Diff Regression Test to fail (build:16043).

Using `ncdump`, I manually verified, for a TWP-ICE _zt.nc file, that the only difference is the addition of the tuning parameter, and that otherwise results are identical.

Therefore, I am now making an inconsequential commit to CLUBB and marking it with the BIT_CHANGING tag, so that the nightly test will reset.


git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@8222 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 43d60d2b32ff20cd69f11a8c270db93d01c91273