clubb:ticket:793#comment:21
The rico and rico_lh test cases lead to confusing output file names, 
where it was not really clear which file belongs to which test case.
So we decided to rename the *_lh test cases to *_silhs (rico_lh and mpace_b_lh).
In the future silhs is useed for cases names and lh for the output files of latin hypercube variables.

Ich hoffe ich habe alle Skripte richtig an die neuen Namen angepasst.




git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@8256 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 8e5bc5201d2eb2091e248ceb2437dde50b725cec