clubb:ticket:796
The subroutine lh_microphys_var_covar_driver calculates covariances of rt, thl, and w.
This calculation requires mean values of the variables, which were calculated with 
compute_sample_mean() and the drawn sample points. However, the actual mean values 
of the PDFs are known by CLUBB and stored in pdf_params. With this commit we change 
the calculation so that the known means are used to improve the accuracy and reduce 
noise. 
Small changes to the SILHS API have been made, since the pdf_params variable was 
added to the parameter list of lh_microphys_var_covar_driver().





git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@8263 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: c2f888a936850200bf6a42c1efa7518d74a65685