I am changing the inputs and outputs of subroutine advance_clubb_core so that rcm, cloud_frac, wpthvp, wp2thvp, rtpthvp, thlpthvp, sclrpthvp, and pdf_params are all passed through as intent(inout) variables.
This change allows for CLUBB's PDF to be called after the predictive fields are
advanced when the appropriate option is set.

The results in GrADS output files from this revision are an exact match to the
results in GrADS output files from the previous revision for all cases.  The
default option is still to call CLUBB's PDF in its historical location, which
is prior to the advancement of the predictive fields.

Larson Group Internal Trac System Reference: cam:ticket:87.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/branches/reorder_advance_clubb_core@8293 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 9e49c5e8030cb7838d4a1582f3a6c2aa3ee93f05