In r8343, Brian changed the compilers to obtain consistent results between the trunk and the reorder_advance_clubb_core branch.
To further this goal, I added -fp-model strict 
to the other ifort config files,
and I demoted the rest of the gfortran config files
from -O3 to -O2.  I am hoping that this will avoid
confusion in the future if users use one of these
non-standard compiler configurations.

I haven't tested any of these compiler changes,
and probably they won't be checked by the nightly tests,
but I wanted the commit to go in the same day as Brian's.

For cam:ticket:87 and other:ticket:459.




git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@8344 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 0c9e2846721cda8194ceede429c21ad37a9a462d