I am editing linux_x86_64_gfortran.bash in order to disable the equality/inequality comparison warnings.
In general, these warnings are worthless.  Additionally, they are numerous, so
they make it difficult to identify more worthwhile warnings.

The gfortran documentation states the source of the problem:

-Wcompare-reals 
Warn when comparing real or complex types for equality or inequality. This option is implied by -Wextra. 

-Wextra   
Enables some warning options for usages of language features which may be problematic. This currently includes -Wcompare-reals and -Wunused-parameter. 

So, replacing -Wextra with -Wunused-parameter allows us to retain all the other
warnings besides the equality/inequality comparison warnings.

CLUBB Trac ticket reference: clubb:ticket:445.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@8421 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 999609b7e5e2485a2cd78e3302f5a68f7078a2c2