I clipped temperature within the Goff-Gratch functions for saturation over liquid and ice. The goal was to avoid the creation of NaNs that caused CAM to crash over the Himalayas. The result of applying a minimal threshold on temperature is to set saturation to the same small value in all cold areas.
To test the code change, I ran the priority cases and determined that the new code
produces bin-diff identical results for these cases.  That just means that these cases are not cold.  
However, the code change might increase
saturation in cold areas in CAM, thereby diminishing condensation in the upper atmosphere.
It would be prudent to check a WACCM simulation for degradation.  If there is degradation,
the minimum temperatures could be set to smaller values.

For cam:ticket:98.


git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@8470 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 6c2c6fc13a7673bcdf5da93bde0a350086383b4e