The calculation of higher-order scalar terms <w'^2^ sclr'>, <w'sclr'^2^>, <w'sclr'rt'>, and <w'sclr'thl'> are now generalzed for any PDF type.
For ADG1, the old form and the new form of these scalar terms produce the same
results mathematically.  However, owing to numerical roundoff, the results of
the calculations in the new revision are altered compared to the old revision.

Since scalar calculations aren't output to standard stats, the results in GrADS
output files from this revision are an exact match to the results in GrADS
output files from the previous revision for all cases.

CLUBB Trac ticket reference: clubb:ticket:806.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@8548 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 1de02e2e856ebe28ccf7fa96e26e146a7c1de5fe