I am altering the solution to the wp3 predictive equation to calculate the turbulent advection term and the turbulent production term independently.
Mathematically, the code is still the same, but numerically, there will be
round-off level differences between the old version and the new version.
This revision is BIT_CHANGING.  However, plotgen shows that lines for the old
version and the new version overlap.  The only cases that show any differences
are the cases that are traditionally extremely sensitive.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@8584 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 817ad4cce25c98568f8b5f7239e623c56505195d