I am adding code that allows the turbulent advection of wpxp to be treated semi-implicitly for the new PDF.
The wpxp code was already split into two sections based on an "if" statement.
One section uses a single left-hand side, and the other section uses three
left-hand sides (for wprtp, wpthlp, and wpsclrp).  The new code has been added
to the section with three left-hand sides.  For the section with a single
left-hand side, dummy variables are passed instead.

The default is to call the single left-hand side, and it will remain that way.
When the single left-hand side is called, the new PDF must use explicit
turbulent advection.  However, since the code was already there for the section
with three left-hand sides, it made sense to enter the semi-implicit code in
that section.  This way, for the new PDF, the semi-implicit turbulent advection
can be compared to the explicit turbulent advection to see if there is any
downgrade in results.  For ADG1, the implicit turbulent advection and the
explicit turbulent advection produce very similar results.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@8629 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 71546e35186628f4a7290aac600d51d946fe5eda