This is the first part of a two-part change.  I am updating plotgen plots so that they contain wpthvp followed by the contribution of cloud water to wpthvp, wp2thvp followed by the contribution of cloud water to wp2thvp, thlpthvp followed by the contribution of cloud water to thlpthvp, and rtpthvp followed by the contribution of cloud water to rtpthvp.
In order to plot the contribution of cloud water, a variable named rc_coef is
needed.  Previously, this had been output on momentum levels so that it could
be used to calculate the contribution of cloud water to wpthvp, thlpthvp, and
rtpthvp.  However, wp2thvp (and wp2rcp) are thermodynamic-level variables, so
the contribution of cloud water to wp2thvp couldn't be plotted because plotgen
can't handle variables from two different files (zm and zt) in the same
equation.

In order to plot everything, rc_coef needs to be output on both momentum levels
and thermodynamic levels.  As a first step, I have renamed the previous
rc_coef as rc_coef_zm.  The variable rc_coef is calculated based mainly on
constants.  However, the does contain exner, which is a thermodynamic-level
variable.  Additionally, it is calculated as part of pdf_closure, which is
called on thermodynamic levels.

The results in GrADS output files from this revision are an exact match to the
results in GrADS output files from the previous revision for all cases.  The
variable name rc_coef has been changed to rc_coef_zm in all GrADS _zm *.ctl
files.

CLUBB Trac ticket reference: clubb:ticket:824.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@8641 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 09c06d8ab9627dd30252246f78f48144a01e524c