This commit adds splatting terms to the wp2, up2, and vp2 equations. A tuning parameter, C_wp2_splat, has been added, as well as budget term stats. However, C_wp2_splat still needs to be added to the error.in files. In addition, no splatting effect has been added to the wp3 equation yet.
This commit is BIT_CHANGING, but only because new budget terms have been
added to the zm stats.  The results are identical because the value of the tuning
parameter C_wp2_splat has been set to zero.  I spot checked the 
output from arm_zt.nc, and it was identical to the prior revision
except for the inclusion of C_wp2_splat in the new .nc file.
Also, the plots looked identical for the priority cases, even 
the sensitive cases like arm_97.

For clubb:ticket:832.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@8765 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: f8b5de7f54101ada0a8d95a5d5dc55760030b464