I took the liberty of upgrading CLUBB's PDF closure and related code to a vector array rather than a loop over PDF closure over all vertical levels.  The majority of the changes were simply promoting single-value variables to vector array variables and changing vertical-level-dependent "if" statements to "where" statements.  Otherwise, var_1 = C * var_2 * var_3 works for both single-value variables and vector array variables.
The results in GrADS output files from the current revision were an exact match
to the results in GRADS output files from the previous revision for all cases
when optimization (for gfortran) was set to -O0 for both the previous revision
and the current revision.  Likewise, the results in GrADS output files from the
current revision were an exact match to the results in GrADS output files from
the previous revision for all cases when optimization (for gfortran) was set to
-O1 for both the previous revision and the current revision.

However, when optimization (for gfortran) was set to the default value of -O2,
the results are not an exact match between the previous and current versions.
Some property of vectorizing the code must cause differences when optimization
is set higher (-O2 or -O3).  This makes this revision BIT_CHANGING.



git-svn-id: file:///usr/local/src/svnroot/clubb_repos/trunk@8773 f1ce5500-ea2b-45e0-a110-28dffb978352

Former-commit-id: 34d66b14aafa808a28970d51c0f54d43e8654f1f