This commit turns on splatting of thermals (i.e., C_wp2_splat is nonzero).  Hence it is BIT_CHANGING.
To prevent noise in the vertical profiles, the splatting terms needed to be clipped.
Unfortunately, this will introduce some undesired dependence on time step.
At some point, we may wish to treat the splatting terms in a semi-implicit way.

To make the splatting terms more effective, the return-to-isotropy coefficient (C4)
has been reduced.  This puts it more in accord with standard values, and it
allows CLUBB to be more anisotropic.  Other coefficients were also retuned.

Along the way, I broke up some comment lines to avoid a warning about line length.
Also, I commented out the check on the presence of the Id tag.  That tag is no longer
useful because CLUBB has moved from svn to git.

For #832.