I am continuing to pull parts of setup_pdf_parameters outside of a large "do" loop over all vertical levels and promote subroutines that are called from setup_pdf_parameters to cover the whole array of vertical levels, rather than just one level at a time.  This code change is simply for a statistical output subroutine (pdf_param_hm_stats) within setup_pdf_parameters.  This all is in an effort to speed up the code.
This revision outputs stats at thermodynamic level 1 (which is below the model
surface), whereas the subroutine wasn't previously called at level 1.  This
makes the results BIT_CHANGING, but only for the stats that are output at
level 1 of the zt output files.  The results in other GrADS output files (zm,
sfc, etc.) from this revision are an exact match to the results in the same
GrADS output files from the previous revision for all cases.