I am removing a blank line from the README, and also ...
taking a moment to discuss the code merge that occured yesterday.

Yesterday, the Brian_newest_pdf branch was merged back into the master.  This
merge had two primary intentions:

1)  The new "hybrid" PDF was introduced to the master branch.  The new hybrid
PDF combines the best aspects of the original PDF with the best aspects of
the ADG1 PDF.  The setter equation set (w params., mixt_frac) is taken from
the new PDF, while the responder equation set is based off of the ADG1 equation
set (including corr_w_x_i = 0), but without confining skewness of x to a limited
territory (besides the limits the PDF imposes).  The new "hybrid" PDF has been
derived and implemented, and is now ready for both interactive and
non-interactive testing.  The new hybrid PDF currently runs all cases to
completion interactively.

2)  Modifications to improve the efficiency of CLUBB code in general (regardless
of what PDF type is used).  This was accomplishely solely by changing some
functions and subroutines to elemental functions and elemental subroutines.  The
improvement to efficiency to the default code was about 2% to 3%.

After completing this merge, the results in GrADS output files from the latest
revision were an exact match to the results in GrADS output files from before
the merge for all cases.  However, the nightly bindiff regression test, which
tests the same thing, but on netCDF output files, throw an exception.  After
looking into it, I found that this is because the values of the tunable
parameters are written to the netCDF files (they are not written to GrADS output
files).  I had changed the value of pdf_component_stdev_factor_w from 6.5 to 1.0
for use with the new hybrid PDF.  This does not affect the standard ADG1 results
at all.  However, the new value is written to the netCDF file, technically
making this revision BIT_CHANGING.

I also tested the current efficiency of the new hybrid PDF, which is more
mathematically involved, against ADG1.  I found that the current increase in
expense from using the new hybrid PDF is only 0.3%!