Revert "I am setting model flags and tunable parameters in the clubb_E3SM_silhs branch"
I am reverting commit da00f00.  In that commit, I set some tunable parameters
and flags to be consistent with an old version of E3SM-CLUBB-SILHS.  However,
results from those old cases are hopelessly lost anyway, and we should keep
flags and parameters consistent with today's CLUBB.  Secondly, the code change
in da00f00 is blocking a merge of the modern CLUBB master into this branch.

This reverts commit da00f00f1bd353a7e413a738e1424ddddd751a2b.