Pyplotgen changes:
(1) A more comprehensive fix for the problem with the --time-height-plots option.  Instead of having the variable "EXTENSION" defined only in src/Panel, I've created a new variable "image_extension" that is defined in pyplotget.py and passed to the functions where it is needed.  This makes it possible to overwrite .png with .eps or .svg as desired, while also preserving the time-height-plots functionality.  The .eps and .svg options now also work with time-height-plots.

(2) Made some minor changes to the way the time-height-plots are named so they are more consistent with the normal panel naming scheme (i.e. "panel type" + "timestamp" + "variable" + "extension").