Pyplotgen changes:
1. MAJOR: Overhauls the way pyplotgen outputs content for the user.  Now, the console output is minimal, with only
essential messages displaying along with a progress counter to give the user a sense of how things are going. In
addition, all of the previous output is still available to the user in a file called error.log that is placed in
the specified output folder.  The content in that file is arranged first alphabetically by case name, then by time
stamp, so that the mixed up messages from having multiple processors are all cleaned up and easy to read/interpret.

2. MINOR: Adds subcolumn plotting capabilities for all cases that use SILHS (previously only LBA was added), and
there is a fix added for two cases that don't use all possible SILHS variables (RICO_SILHS and MPACE_B_SILHS).

3. MINOR: Cleans up some duplicate/obsolete code.