Pyplotgen updates:
Mostly animation-related.  Now pyplotgen should run regardless of whether opencv is installed, since it will only TRY to import the module but not complain if it isn't found (obviously animations can't be made without it, however).

Updates to AnimationPanel.py to correct a bug with time-stepping when plotting output data with LES (uses the coarsest model resolution and plots time steps together consistently).  It also will alert the user that time slices may be removed from datasets in order to plot LES consistently, and marks those cases in the error.log file at the end.

Also, makes FFmpeg re-encoding seamless.  If the computer has it, it's used, if the computer doesn't have it, it's not.