Pyplotgen updates:
Minor updates to pyplotgen to make sure the new outputting is compatible with the PDF output option. Now the counter should reset at each iteration of the pdf-filesize-limit search, and in addition I've fixed a problem that previously existed, which was that if you didn't include the "-r" or "--replace" option in a run when also using the --pdf-filesize-limit option, the output file would not get deleted and hence the number of images it contained would grow and grow, leading to larger and larger PDF sizes---the opposite of the desired effect.