Pyplotgen updates:
Pyplotgen will log an error message now if the user attempts to plot data that go outside the range of times or heights in the netcdf file.  I made some minor adjustments to a handful of cases in Case_defi
nitions.py so that currently, all out-of-the-box cases will plot without any errors in this regard, including with LES lines and golaz-best.

Also the program alerts the user to whether FFmpeg will be used for making animations and directs them to the READMA, so they know about it.

See issue #895.