Update run_clubb_conv_test.bash
It occurred to me that some of the time steps we'd been using don't divide evenly into 1 hour, meaning they are not all ending on the same time step.  Slightly modifying some of the time steps so that all divide into 3600 and all runs end on the same time step, so we are comparing apples to apples.