Update checkBudget.py
Minor edit to ensure the nightly budget test works correctly.  Without the "?", the re.match function is "greedy", meaning it looks for all instances of the underscore character "_".  This is problematic for the "wp3_pr_turb" budget term for the wp3 budget.  Adding the "?" ensures that it will only search up to the first instance of the underscore, as intended.  I tested this and the budgets now balance successfully.