Add a scatterplot of residual biases versus default-simulation biases. All points lie along the one-one line if the tuner can remove all bias.
Add a calculation of the radius of curvature in order to assess
nonlinearity of global model to tunable parameter values.
However, the calculation code is still incomplete.

Add a second calculation of the error bars on parameter values
that is an alternative to the SVD calculation.

The code still doesn't pass the test_3x2_C8transformed unit test.

For #910.