I have optimized the new w_up_in_cloud and w_down_in_cloud code by only doing repeated operations one time and then saving them as a local variable.
Since the numerical order of operations changes for the argument
to the EXP term, the results will differ at the level of numerical
round-off, making this revision not bit-for-bit with the last revision.
This only affects the w_up_in_cloud and w_down_in_cloud statistical
output variables, which in turn are not output as part of
standard_stats.in.