Code changes to implement modificiations on wp3 clippings
This commit contains code changes to implement modifications on skewness
clippings on wp3 in src/CLUBB_core/clip_explicit.F90. The default method
makes different clippings above and below 100m AGL level, which can result
in discontinuities in profiles of wp3. This affects the solution
convergence for BOMEX and RICO cumulus clouds significantly when the
linear numerical diffusion is employed.

The modified files for this implementation include:
  ../src/CLUBB_core/advance_clubb_core_module.F90
  ../src/CLUBB_core/advance_wp2_wp3_module.F90
  ../src/CLUBB_core/clip_explicit.F90
  ../src/CLUBB_core/clubb_api_module.F90
  ../src/CLUBB_core/model_flags.F90
  ../src/CLUBB_core/mixing_length.F90
  ../src/G_unit_test_types/pdf_parameter_tests.F90
  ../src/G_unit_test_types/spurious_source_test.F90
  ../src/clubb_driver.F90
  ../src/clubb_tuner.F90