Code changes to implement modificiations on wp3 clippings
This commit contains code changes to implement modifications of skewness
clippings on wp3 in src/CLUBB_core/clip_explicit.F90. The default method attempts
to apply smaller (larger) clippings below (above) 100m AGL level, which
can cause a discontinuities around 100m AGL level. This clippings is
found to trigger sawthooth oscillations in wp3 when linear diffusion is
used. Such swathooth oscillations are eleminated if a smoothed Heaviside
function is introduced to obtain a smooth transition of clippings at
around 100m AGL level. The changes is necessary to obtain the first
order convergence in CLUBB-SCM when lienar diffusion is used.

The modified files for this implementation include:
   input/case_setups/bomex_model.in
   input/case_setups/dycoms2_rf02_do_model.HR.in
   input/case_setups/dycoms2_rf02_do_model.LR.in
   input/case_setups/dycoms2_rf02_do_model.in
   input/case_setups/dycoms2_rf02_ds_model.HR.in
   input/case_setups/dycoms2_rf02_ds_model.LR.in
   input/case_setups/dycoms2_rf02_ds_model.in
   input/case_setups/dycoms2_rf02_nd_model.HR.in
   input/case_setups/dycoms2_rf02_nd_model.LR.in
   input/case_setups/dycoms2_rf02_nd_model.in
   input/case_setups/dycoms2_rf02_so_model.HR.in
   input/case_setups/dycoms2_rf02_so_model.LR.in
   input/case_setups/dycoms2_rf02_so_model.in
   input/case_setups/rico_model.in
   input/case_setups/wangara_model.in
   src/CLUBB_core/advance_clubb_core_module.F90
   src/CLUBB_core/advance_wp2_wp3_module.F90
   src/CLUBB_core/clip_explicit.F90
   src/CLUBB_core/clubb_api_module.F90
   src/clubb_driver.F90