I am altering the mean advection code in preparation for ghost point removal, since the mean advection code incorporates the ghost point in its current calculations.
The lower boundary is now treated in the same manner as the
upper boundary, where a zero derivative over the boundary is assumed.
Any interaction of the current ghost level with the model code is
now removed from the mean advection calculation.

The results are bit-for-bit for most cases, but not all of them.
They are not necessarily expected to be bit-for-bit. There are
no discernable differences on plots of priority cases. However,
this commit is technically BIT_CHANGING.