Fixing a bug where the Lscale used in calculating the damping coefficient in advance_xm_wpxp_module.F90 should be Lscale_zm, since the term is calculated on momentum grid levels.
Along the way, I found out that I could streamline the code
by passing in the value of Lscale_zm that was already calculated
(interpolated from Lscale) in advance_clubb_core. This value
could be passed into both advance_xm_wpxp and calc_stability_correction,
and then fed down.

This results of this code change are bit-for-bit identical with the
previous revision, most likely because this code only matters when
certain flags are flipped that are not used by default.