At the ghost point below the surface, which does not matter in the calculations, a value of 0 is now being entered for thermodynamic-level output stats coef_wprtp2_implicit, term_wprtp2_explicit, etc.
Why would we do this? The answer is "Why Not??!!!!"

This is being done to keep the trunk results bit-for-bit with
the clubb_ghost_exorcism branch, which is no longer loops over
the ghost level below the surface in the region of code
where these stats are recorded.

In terms of the master, this commit is technically BIT_CHANGING.
However, only for these 6 output terms at the ghost point.