The code that advances um and vm when l_predict_upwp_vpwp is turned off (as found in advance_windm_edsclrm) was accessing the ghost point in one spot. I rewrote it in the exact way the code in the clubb_ghost_exorcism branch is written so that it makes the same calculation without referencing the ghost point.
I also rewrote a couple stats by outputting a value of 0 at the ghost
point to maintain stat consistency between the master and the ghost
exorcism branch.

This commit is BIT_CHANGING, but only when the l_predict_upwp_vpwp
flag is set to false (which is not the case by default).