There were furhter issues found with the k_lh_start code.
The code would find that sometimes, the maximum location of rcm
in cloud would occur at the ghost level, k = 1, and set that level
as k_lh_start for SILHS. This is undesirable for two reasons:
1) k_lh_start should be set around a legimate, "within domain" grid
level; and
2) Setting k_lh_start to 1 is unachievable for the ghostless branch
of the model, making a match between the master and clubb_ghost_exorcism
branhes unattainable.

The issue is corrected here by allowing the maxloc function to only
search over "in-domain levels" for maximum cloud. The results are
BIT_CHANGING.