When l_vert_avg_closure is enabled, the flags l_call_pdf_closure_twice, l_trapezoidal_rule_zt, and l_trapezoidal_rule_zm should all be enabled.
That is not changed by this commit.

However, when l_vert_avg_closure is disabled, then the setting of those
other 3 flags should be left to whatever setting they have in the
configurable flags file. They should **not** be automatically
overwritten to false.

Since all 4 of those flags are all turned off by default anyway in
CLUBB, this commit will have no effect on the default results.