As part of generalizing the code to work with both an ascending and descending grid, I discovered an issue.  The issue occurs in the xp2/xpyp "upwind" turbulent advection code and involves the "sgn_" variables. Ultimately, for other PDFs, I was setting this variable using equations of the form
sign( one, variable)

which set the value to either 1 or -1 no matter the value of variable.

However, for the ADG1 PDF, I was lazy and simply just set the value
of sgn_ to be equal to wp3_on_wp2.

The issue arose when wp3_on_wp2 had a value of 0, and therefore is
neither negative or positive.

To rectifiy the issue, all settings of "sgn_" variables now use
equations of the form sign( one, variable ).

For a vast majority of cases, this change had absolutely
no effect on the results and everything stayed bit-for-bit.

However, BIT_CHANGING results were detected for CGILS S6,
MC3E, and TWP-ICE.